requirements.txt
beautifulsoup4
chromadb
fastapi
langchain
openai_whisper
pydantic==1.10.9
python-dotenv
pytube
PyYAML
Requests
scrapetube
slack_bolt
tiktoken
whisper
openai
